`xcp-rrdd` used to ignore timestamps of measurement times provided by
the RRDD plugins, and always calculated change in all values against a
single timestamp captured at the time of reading on the server. This can
potentially trigger a case where the time that passed between
measurements is not equal to the time that passed between readings (or
other measurements), and values would be calculated incorrectly. This
can be triggered on a host with 100% CPU load, where `cpu0-P0` derived
metric, for example, would often be > 1.0.

This PR reworks `xcp-rrdd` and the plugin infrastructure to calculate
values based on the actual length of time passed since the last value
for this particular metric was recorded. This required extensive
architectural changes, explained in detail in commits themselves - *so
best reviewed by commit*.

I've tested this over the weekend, and haven't registered any spikes in
metrics on a fully-loaded host. Will run BST+BVT as well, since RRDD is
at the center of a lot of other things and is quite fragile.

+++ ocaml/xcp-rrdd/lib/transport/base/rrd_protocol.ml
@@ -26,7 +26,7 @@ exception Payload_too_large
 
 exception Read_error
 
-type payload = {timestamp: int64; datasources: (Rrd.ds_owner * Ds.ds) list}
+type payload = {timestamp: float; datasources: (Rrd.ds_owner * Ds.ds) list}
 
 type protocol = {
     make_payload_reader: unit -> Cstruct.t -> payload
